requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-filter==2.4.0
django-jet==1.0.8
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
jet==1.0
Pillow==8.0.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
psycopg2==2.6.2
